Rename ProdVsCost to TempProdVsCost and colocate assets
